bug fix launch into cli
