+++ src/controller/gController/__main__.py
@@ -1,3 +1,3 @@
 from gview import *
 
-main()
+from .GController import *
